0-value integers(or numbers) read as null issue resolved;
If in an int column (for example) a null value was inserted, that value
was read as null which was not correct.
If there were more columns after that one, everything was read as null,
even if values were actually there.
This issue happened for all types of numbers.

Conflicts:
	src/Protocol/Response/DataStream.php